requirements.txt
fastapi~=0.110.1
SQLAlchemy~=2.0.29
tqdm~=4.66.2
pymysql
uvicorn
tortoise-orm